Move cube_to_image and cube_to_spec and their test to cube package
